% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{appendRDS}
\alias{appendRDS}
\title{(Internal) Append RDS files}
\usage{
appendRDS(object, file, wd)
}
\arguments{
\item{object}{The object to save or append}

\item{file}{The name of the file to be saved to be saved}

\item{wd}{The directory to check}
}
\description{
(Internal) Append RDS files
}
