% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{auto_label_unlabeled_values}
\alias{auto_label_unlabeled_values}
\title{(Internal) Label unlabelled values in imported SAS files}
\usage{
auto_label_unlabeled_values(lbl_vector, wd = wd, x = x, varname)
}
\arguments{
\item{lbl_vector}{A vector with labels}

\item{wd}{Working directory for files}

\item{x}{NCSA table name (sas file name)}

\item{varname}{Variable name or label}
}
\description{
(Internal) Label unlabelled values in imported SAS files
}
