% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cases.R
\name{bicyclist}
\alias{bicyclist}
\title{(Internal) Find crashes involving bicyclists}
\usage{
bicyclist(FARS)
}
\arguments{
\item{FARS}{The FARS data object to be searched.}
}
\description{
These internal functions take the FARS object created by use_fars and look
    for various cases, such as distracted or drowsy drivers.
}
