% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_internet_connection}
\alias{check_internet_connection}
\title{(Internal) Check internet connection}
\usage{
check_internet_connection()
}
\value{
Logical indicating whether internet connection is available
}
\description{
Test if internet connection is available by attempting to reach a reliable host.
This function is used to gracefully handle cases where internet resources
are not available.
}
