% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\title{Get FARS data}
\usage{
counts(df, what, when, interval, where, who, why)
}
\description{
This function is a wrapper for \code{download_fars} and \code{prep_fars}.
    It downloads raw files from NHTSA and produces analysis-ready data files.
}
\details{
Prepared data files are stored in \code{save_dir}/FARS data/prepared/
}
\examples{
get_fars(c("2019", "2020"))
get_fars(2016:2020, "Virginia")
get_fars(2020, "NC")
}
\seealso{
\code{download_fars} \code{prep_fars}
}
