% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cases.R
\name{driver_age}
\alias{driver_age}
\title{(Internal) Find crashes involving drivers of a given age}
\usage{
driver_age(df, age_min, age_max)
}
\arguments{
\item{df}{The FARS or GESCRSS data object to be searched.}

\item{age_min}{Lower bound on driver age (inclusive).}

\item{age_max}{Upper bound on driver age (inclusive).}
}
\description{
These internal functions take the FARS object created by use_fars and look
    for various cases, such as distracted or drowsy drivers.
}
