% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fars_1920_va_flat}
\alias{fars_1920_va_flat}
\title{FARS Data Sample (Virginia, 2019-2020)}
\format{
A dataframe with 3,496 rows and 30 variables.
}
\usage{
fars_1920_va_flat
}
\description{
A toy dataset taken from the 'flat' element of a FARS data object (created by rfars::get_fars(years=2019:2020, states="VA")).
   Less commonly used variables have been omitted to minimize memory requirements.
}
\details{
See the fars_codebook for more information about each variable.
}
\seealso{
[rfars::get_fars()], "fars_codebook"
}
\keyword{datasets}
