% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fars_data_changes}
\alias{fars_data_changes}
\title{Changes in FARS Data Elements by Data File and Year}
\format{
A data frame with 46 rows and 480 columns.
}
\source{
\url{https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/813254}
}
\usage{
fars_data_changes
}
\description{
A dataset describing major changes to the FARS data system over time.
}
\details{
See Appendix F of the \href{https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/813254}{2020 Analytical User’s Manual} for more information.
}
\keyword{datasets}
