% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fars_data_structure}
\alias{fars_data_structure}
\title{FARS data structure}
\format{
A data frame with 27 rows and 4 columns.
\describe{
  \item{tablename}{the cleaned name of the data file}
  \item{structure}{either one or multiple, indicating the number of rows per entity}
  \item{level}{the entity level (crash, vehicle, or person) or the data file}
  \item{year_created}{the first year that the data file was in use}
}
}
\source{
Page 19 of the \href{https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/813254}{2020 Analytical User’s Manual}
}
\usage{
fars_data_structure
}
\description{
A dataset describing the structure and level of each raw FARS data file.
}
\keyword{datasets}
