% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the car fatalities in a US state and in a specific year.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer value corresponding to a US state.}

\item{year}{A numeric vector representing the years in a time period.}
}
\description{
It plots, in a map, the car fatalities occurred in a given US state and in a
specific year using the data reported by the US National Highway Traffic
Safety Administration in the Fatality Analysis Reporting System.
}
\section{Warning}{

   If the \code{state.num} parameter is not a valid state number or the
   information specified in the \code{year} parameter does not exist, the
   function will throw an error.
}

\examples{
# We set the working directory where the compressed files are stored within
# the package.
WD <- setwd(system.file("extdata", package = "rfars"))
fars_map_state(12, 2015)
setwd(WD)
}
