% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a delimited file into a tibble object.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the compressed file which data are to be read
from (a string value).
\emph{If the file does not exist in the working directory},
\bold{you should specify the whole path}.}
}
\value{
A tibble object with the contents of the filename.
}
\description{
The function \code{fars_read} reads the contents of a compressed file in .bz2
format and returns them in a
\href{https://blog.rstudio.org/2016/03/24/tibble-1-0-0/}{tibble} object.
}
\details{
The function uses the algorithm of the \code{\link[readr]{read_csv}}
   function.
}
\section{Warning}{

   Provided that the file does not exist, the function will throw an error.
}

\examples{
# We set the working directory where the compressed files are stored within
# the package.
WD <- setwd(system.file("extdata", package = "rfars"))
fars_read("accident_2013.csv.bz2")
setwd(WD)
}
