% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Creates a summary table with the counts of car fatalities.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector representing the years of a time period.}
}
\value{
A \href{https://blog.rstudio.org/2016/03/24/tibble-1-0-0/}{tibble}
   object with the tallies of car fatalities by month for each year in the
   specified period.
}
\description{
It creates a summary table with the tallies of the car fatalities reported
by the US National Highway Traffic Safety Administration in the Fatality
Analysis Reporting System for a specific period.
}
\section{Warning}{

   Provided that a file does not exist for a specific year, the function will
   throw a warning that the information does not exist.
}

\examples{
# Summary with the tallies of car fatalities in 2013.
fars_summarize_years(2013)
# Summary with the tallies of car fatalities in 2013 and 2014.
fars_summarize_years(c(2013, 2014))
}
