% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fars_varnames}
\alias{fars_varnames}
\title{FARS Variable Names}
\format{
A data frame with 468 rows and 4 columns
}
\source{
\url{https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/813254}
}
\usage{
fars_varnames
}
\description{
A dataset that translates machine-readable variable names to friendly names
}
\details{
See \href{https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/813254}{2020 Analytical User’s Manual} for more information.

\describe{
  \item{table}{the cleaned name of the data file}
  \item{original}{the original variable name}
  \item{friendly}{human-readable (friendly) version of the variable name}
  \item{original_clean}{the cleaned name of the variable}
 }
}
\keyword{datasets}
