% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_relations}
\alias{geo_relations}
\title{Synonym table for various geographical scales}
\format{
A data frame with 3,142 rows and 6 variables:
\describe{
  \item{FIPS}{5-digit FIPS code composed of \code{state_fips} and \code{county_fips}}
  \item{state_fips}{2-digit FIPS code indicating a state}
  \item{county_fips}{3-digit FIPS code indicating a county within a state}
  \item{state_abbr}{2-character, capitalized state abbreviations}
  \item{state_name}{fully spelled and case-sensitive state names}
  \item{county_name}{fully spelled and case-sensitive state names}
}
}
\source{
\url{https://www.census.gov/geographies/reference-files/2015/demo/popest/2015-fips.html}
}
\usage{
geo_relations
}
\description{
A dataset providing different ways to refer to states and counties.
}
\keyword{datasets}
