% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gescrss_1920_mw_flat}
\alias{gescrss_1920_mw_flat}
\title{GESCRSS Data Sample (Midwest, 2019-2020)}
\format{
A dataframe with 44,956 rows and 29 variables.
}
\usage{
gescrss_1920_mw_flat
}
\description{
A toy dataset taken from the 'flat' element of a GESCRSS data object (created by rfars::get_gescrss(years=2019:2020, regions="mw")).
   Less commonly used variables have been omitted to minimize memory requirements.
}
\details{
See the gescrss_codebook for more information about each variable.
}
\seealso{
[rfars::get_gescrss()], "gescrss_codebook"
}
\keyword{datasets}
