% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_sas_attrs}
\alias{get_sas_attrs}
\title{(Internal) Check SAS attributes}
\usage{
get_sas_attrs(data)
}
\arguments{
\item{data}{An object produced by haven::read_sas()}
}
\description{
(Internal) Check SAS attributes
}
