% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{import_multi}
\alias{import_multi}
\title{(Internal) Import the multi_ files}
\usage{
import_multi(filename, where)
}
\arguments{
\item{filename}{The filename (e.g. "multi_acc.csv") to be imported}

\item{where}{The directory to search within}
}
\description{
An internal function that imports the multi_ files
}
