% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_all_numeric}
\alias{make_all_numeric}
\title{(Internal) Make id and year numeric}
\usage{
make_all_numeric(df)
}
\arguments{
\item{df}{The input dataframe}
}
\description{
(Internal) Make id and year numeric
}
