% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_id}
\alias{make_id}
\title{(Internal) Generate an ID variable}
\usage{
make_id(df)
}
\arguments{
\item{df}{The dataframe from which to make the id}
}
\description{
(Internal) Generate an ID variable
}
