% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_sas_format}
\alias{parse_sas_format}
\title{(Internal) Parse formats.sas instead of using a .sas7bcat file}
\usage{
parse_sas_format(file_path)
}
\arguments{
\item{file_path}{The path of the formats.sas file}
}
\description{
(Internal) Parse formats.sas instead of using a .sas7bcat file
}
