% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_fars.R
\name{prep_fars}
\alias{prep_fars}
\title{Prepare FARS data}
\usage{
prep_fars(raw_dir = getwd(), states = NULL)
}
\arguments{
\item{raw_dir}{Directory where raw files are currently saved.}

\item{states}{(Optional) states to keep. Leave as NULL to keep all states.
Can be specified as full state name (e.g. "Virginia"), abbreviation ("VA"),
or FIPS code (51).}
}
\value{
Produces four files for each year: yyyy_flat.csv, yyyy_multi_acc.csv,
    yyyy_multi_veh.csv, and yyyy_multi_per.csv.
}
\description{
Combine raw files to create analysis-ready FARS data files.
}
\details{
Flat files are wide-formatted and presented at the person level.
    All \emph{crashes} involve at least one motor \emph{vehicle}, each of
    which may contain one or multiple \emph{people}. These are the three
    entities of crash data. The flat files therefore repeat some data elements
    across multiple rows. Please conduct your analysis with your entity in mind. \cr \cr
    Some data elements can include multiple values for any data level
    (e.g., multiple weather conditions corresponding to the crash, or multiple
    crash factors related to vehicle or person). All of these elements have been
    collected in the yyyy_multi_[acc/veh/per].csv files in long format.
    These files contain crash, vehicle, and person identifiers, and two
    variables labelled \code{name} and \code{value}. These correspond to
    variable names from the raw data files and the corresponding values,
    respectively. \cr \cr
    The flat files contain one row per person (which may result in multiple
    rows for associated vehicles and crashes), but the multi files can
    contain a variable number of rows for any crash entity. \cr \cr
    Consult the \href{https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/813254}{Analytical User’s Manual}
    for more information.
}
\examples{
prep_fars()
prep_fars("Virginia")
prep_fars("NC")
}
