% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_fars_2015.R
\name{prep_fars_2015}
\alias{prep_fars_2015}
\title{Handle yyyy data preparation}
\usage{
prep_fars_2015(y, wd, rawfiles, prepared_dir, geo_filtered)
}
\arguments{
\item{y}{year, to be passed from \code{prep_fars}}

\item{wd}{working directory, , to be passed from \code{prep_fars}}

\item{rawfiles}{dataframe translating filenames into standard terms,
to be passed from \code{prep_fars}}

\item{prepared_dir}{the location where prepared files will be saved,
to be passed from \code{prep_fars}}

\item{geo_filtered}{dataframe of filtered geo-identifiers, to be passed
from \code{prep_fars}}
}
\value{
Produces four files for each year: yyyy_flat.csv, yyyy_multi_acc.csv,
    yyyy_multi_veh.csv, and yyyy_multi_per.csv
}
\description{
Handle yyyy data preparation
}
