% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gescrss.R
\name{prep_gescrss}
\alias{prep_gescrss}
\title{Prepare downloaded GES/CRSS files for use}
\usage{
prep_gescrss(y, wd, rawfiles, prepared_dir, regions)
}
\arguments{
\item{y}{year, to be passed from \code{prep_gescrss}}

\item{wd}{working directory, , to be passed from \code{prep_gescrss}}

\item{rawfiles}{dataframe translating filenames into standard terms,
to be passed from \code{prep_gescrss}}

\item{prepared_dir}{the location where prepared files will be saved,
to be passed from \code{prep_gescrss}}

\item{regions}{(Optional) Inherits from get_gescrss()}
}
\value{
Produces six files: yyyy_flat.rds, yyyy_multi_acc.rds,
    yyyy_multi_veh.rds, yyyy_multi_per.rds, yyyy_events.rds, and codebook.rds
}
\description{
Prepare downloaded GES/CRSS files for use
}
