% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_basic_csv.R
\name{read_basic_csv}
\alias{read_basic_csv}
\title{Internal function that takes care of basic CSV reading}
\usage{
read_basic_csv(x, wd = wd, rawfiles = rawfiles)
}
\description{
Internal function that takes care of basic CSV reading
}
