% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_basic_sas.R
\name{read_basic_sas}
\alias{read_basic_sas}
\title{Internal function that takes care of basic SAS importing}
\usage{
read_basic_sas(x, wd, rawfiles)
}
\arguments{
\item{x}{The cleaned name of the data table (CSV).}

\item{wd}{The working directory for these files}

\item{rawfiles}{The data frame connecting raw filenames to cleaned ones.}
}
\description{
Internal function that takes care of basic SAS importing
}
