% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rm_cols.g}
\alias{rm_cols.g}
\title{(Internal) rm_cols.g}
\usage{
rm_cols.g(df, a, b)
}
\arguments{
\item{df}{The input data frame.}

\item{a}{The original, non-imputed variable.}

\item{b}{The imputed variable (often with an _im suffix).}
}
\description{
An internal function that removes variables that are unnecessarily duplicated across GES/CRSS tables.
}
