% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rural_pct}
\alias{rural_pct}
\title{County-level estimation of rurality}
\format{
A data frame with 3,142 rows and 5 variables:
\describe{
  \item{FIPS}{5-digit FIPS code composed of \code{state_fips} and \code{county_fips}}
  \item{pop2010_total}{total population of a given county}
  \item{pop2010_urban}{population living in urban areas}
  \item{pop2010_rural}{population living in rural areas}
  \item{pop2010_rural_pct}{the percentage (0-100) of a county's population
      living in rural areas, hence, the \emph{percent-rural}}
}
}
\source{
\url{https://www.census.gov/programs-surveys/geography/guidance/geo-areas/urban-rural.html}
}
\usage{
rural_pct
}
\description{
A dataset provided by US Census describing urban and rural populations in
    all US counties based on the 2010 Census.
}
\keyword{datasets}
