% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_fars.R
\name{use_fars}
\alias{use_fars}
\title{Use FARS data files}
\usage{
use_fars(prepared_dir = getwd(), years = NULL)
}
\arguments{
\item{prepared_dir}{Directory where prepared files are currently saved.}

\item{years}{(Optional) Years to use.}
}
\value{
Returns either a single data frame (if \code{multi=NULL}) or a list
    containing the combined flat file data frame and specified \code{multi}
    files.
}
\description{
Combine multiple years of prepared FARS data stored in CSV files and bring
    into the current environment.
}
\examples{
\dontrun{
myData <- use_fars()
}
}
\seealso{
\code{download_fars()} \code{prep_fars()} \code{get_fars()}
}
