% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gescrss.R
\name{use_gescrss}
\alias{use_gescrss}
\title{(Internal) Use GESCRSS data files}
\usage{
use_gescrss(dir, prepared_dir, cache)
}
\arguments{
\item{dir}{Inherits from get_gescrss().}

\item{prepared_dir}{Inherits from get_gescrss().}

\item{cache}{Inherits from get_gescrss().}
}
\value{
Returns an object of class 'GESCRSS' which is a list of six tibbles:
    flat, multi_acc, multi_veh, multi_per, events, and codebook.
}
\description{
Compile multiple years of prepared GESCRSS data.
}
