% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{use_imp}
\alias{use_imp}
\title{(Internal) use_imp}
\usage{
use_imp(df, original, imputed, show = FALSE)
}
\arguments{
\item{df}{The input data frame.}

\item{original}{The original, non-imputed variable.}

\item{imputed}{The imputed variable (often with an _im suffix).}

\item{show}{Logical (FALSE by default) Show differences between original and imputed values.}
}
\description{
An internal function that uses imputed variables (present in many GES/CRSS tables)
}
