% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usenames.R
\name{usenames}
\alias{usenames}
\title{(Internal) Use decoded variables instead of encoded ones}
\usage{
usenames(df)
}
\arguments{
\item{df}{Data frame with both versions of some variables.}
}
\value{
A data frame with the encoded variables replaced with decoded versions.
}
\description{
The raw data files include two versions of many data elements: an encoded one
    (using numbers to indicate characteristics such as injury severity,
    relation to the roadway, race, etc.) and another that has already been
    decoded. These variables are labelled according to the convention: x and
    xname, where the latter is the decoded version. This internal function
    goes through a given data frame, removing the decoded versions and
    renaming the encoded ones to remove the 'name' suffix.
}
\seealso{
\code{prep_fars()}
}
