% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{validate_states}
\alias{validate_states}
\title{(Internal) Validate user-provided list of states}
\usage{
validate_states(states)
}
\arguments{
\item{states}{States specified in get_fars, prep_fars, or counts}
}
\description{
(Internal) Validate user-provided list of states
}
