% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_years.R
\name{validate_years}
\alias{validate_years}
\title{(Internal) Validate user-provided list of states}
\usage{
validate_years(years)
}
\arguments{
\item{years}{Years specified in download_fars, get_fars, prep_fars, or counts}
}
\description{
(Internal) Validate user-provided list of states
}
