% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{counts}
\alias{counts}
\title{Generate a variety of counts from FARS data}
\usage{
counts(
  FARS,
  what = "crashes",
  years = NULL,
  interval = "year",
  where = NULL,
  who = NULL,
  involved = NULL,
  filterOnly = FALSE
)
}
\arguments{
\item{FARS}{The input FARS data with flat and multi components.}

\item{what}{What to count: crashes, fatalities, people involved.}

\item{years}{The years over which to count.}

\item{interval}{The interval in which to count: months or years.}

\item{where}{Where to count: can specify rural/urban and/or state (e.g.,
where = "rural Virginia", where = "rural", where = "North Carolina")}

\item{who}{The type of person to count: driver, passenger, pedestrian,
bicyclist, or motorcyclist.}

\item{involved}{Factors involved with the crash. Can be any of: distracted
driver, drowsy driver, police pursuit, motorcycle, pedalcyclist,
bicyclist, pedestrian, pedbike, young driver, older driver, speeding,
alcohol, drugs, hit and run, roadway departure, rollover, or large
trucks.}

\item{filterOnly}{Logical, whether to only filter or reduce to counts.}
}
\description{
Generate a variety of counts from FARS data
}
\details{
...
}
\examples{
\dontrun{
get_fars(years = 2016:2020, states="Virginia") \%>\%
  use_fars()
  counts(
    what = "fatalities",
    years = 2015:2020,
    who = c("bicyclists", "pedestrians"),
    where = "urban"
    ) \%>\%
  ggplot(aes(x=date, y=n, label=scales::comma(n))) + geom_col() + geom_label()
  }
}
