% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_fars.R
\name{download_fars}
\alias{download_fars}
\title{(Internal) Download FARS data files}
\usage{
download_fars(years, dest_raw, dest_prepd)
}
\arguments{
\item{years}{Years to be downloaded, in yyyy (character or numeric formats)}

\item{dest_raw}{Directory to store raw CSV files}

\item{dest_prepd}{Directory to store prepared CSV file}
}
\value{
Nothing, called for side effects.
}
\description{
Download files from NHTSA, unzip, and prepare them.
}
\details{
Raw files are downloaded from \href{https://www.nhtsa.gov/file-downloads?p=nhtsa/downloads/FARS/}{NHTSA}.
}
