% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot fatal injures locations on state map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{integer with state number required}

\item{year}{year required}
}
\value{
This function returns state map with plotted points of fatal accidents or show message that there is nothing to be plotted
}
\description{
This is a function for plotting fatal accidents on state map according to FARS data with specified \code{state.num}
and \code{year}. In case of wrong \code{state.num} function will throw "invalid STATE number" error.
}
\examples{
fars_map_state(10, 2015)

}
