% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw a state map with fars points}
\usage{
fars_map_state(state.num, year)
}
\description{
This function creates a state map with points representing accidents
overlaid on the plotted map.
}
