% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Vizualize traffic accidents happening in given states during specific years}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the code number of a state, which can be in any format, but should be singular}

\item{year}{a year in which to search for traffic accidents, which can be in any format, but
should be singular}
}
\value{
a plot with the locations of accidents in a given state in a given year, mapped over the
       outline of the state. if a state number does not exist or if the given state has no
       accidents in that year, it returns an error message.
}
\description{
This function creates a graph of the traffic accidents that have happened
         during given years in given states, plotted by location over an outline of the state.
}
\examples{
\dontrun{
fars_map_state(state.num=1,year=2013)
fars_map_state(1,2013)
}

}
