% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string that specify file to read}
}
\value{
This function returns file content in "tbl_df" format
}
\description{
This is a function to open file with specified \code{filename} argument.
In case of wrong path it will show error "file \code{filename} does not exist.
}
\examples{
fars_read("data/file.txt")

}
