% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a delimited file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the path to a file.}
}
\value{
This function returns a tibble of the specified data file. Using a
   missing or invalid value for a file or path name will result in an error.
}
\description{
This function reads a file in a tibble format and creates a data frame from
it. In the data frame, cases correspond to rows and variables to columns in
the file.
}
\details{
First, the function will check if the file named by the argument
   exists. If it does not, an error will be thrown indicating the specified
   file does not exist. If the file name is valid, the function reads the
   data and converts it into tibble format.
}
\examples{
fars_read("accident_2013.csv.bz2")

}
