% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Load csv data into R in data frame tbl format}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character string for the file name input of the .csv file}
}
\value{
this function either returns a data frame tbl of the .csv data, or an error,
       if a file for the given year does not exist
}
\description{
This function uses the read_csv function from the readr package to load a given
         .csv file into R, and then converts the imported data into the data frame tbl structure.
         If the file does not exist, the function will stop and return an error.
}
\examples{
\dontrun{
fars_read(filename="accident_2013.csv.bz2")
fars_read("accident_2014.csv.bz2")
}

}
