% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Gain month/year combinations for FARS data read}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{years required}
}
\value{
This function returns a list of data frames with month/year combination
}
\description{
This is the function to read FARS files for specified \code{years}
and get a list of month/year data frames when succeed. In case of incorrect year, function will throw warning "invalid year: \code{year}".
}
\examples{
fars_read_years(2013:2015)
fars_read_years(c(2013, 2015))

}
