% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Load and tidy all accident data for given years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{the years to find files for to load data into R. can be a vector or a single year}
}
\value{
a list of data frame tbls, for given years, each with month and year columns, or an error,
       if a file for the given year does not exist.
}
\description{
For each year input, this function first calls the make_filename function to generate
          the files name for the data on accidents in that year, then loads the generated filename
          into R using the fars_read function, and finally reduces the data to just month and
          years columns.  If a file with a given year does not exist, it will return an error.
}
\examples{
\dontrun{
fars_read_years(years=c(2013,2014,2015))
fars_read_years(c(2013,2014,2015))
}

}
