% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Read data for specified year(s) and combine into a data frame}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector specifying the year(s) (YYYY format).}
}
\value{
This function returns a tibble for the data and selects the month and
   year columns.
}
\description{
This function reads a file for a specified year, uses the year to create a
column name, selects the year and month columns, and converts the list into
a single data frame in tibble format.
}
\details{
This function takes the year input, reads in the data using
   \link{fars_read_years}, and stores the resulting data frames into a list.
   It then combines those data frames into a single tibble, the number of
   cases there are of each.
}
\examples{
fars_read_years(2013)
fars_read_years(c(2014, 2015))

}
