% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the number of accidents per month in a table of years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{the years to find files for to load data into R. can be a vector or a single year}
}
\value{
a data frame of accident counts by month, spread according to the years factor
}
\description{
This function first calls the fars_read_years function to read in data from accident
          records for a given set of years.  It then compiles all of the data frame tbls it's
          read in into a single data frame table by binding all the rows.  It then creates a
          a summary count of accidents per month, per year, and spreads the data frame tbl into
          a data frame in the form of a table of months vs years.
}
\examples{
\dontrun{
fars_summarize_years(years=c(2013,2014,2015))
fars_summarize_years(c(2013,2014,2015))
}

}
