% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo_relations}
\alias{geo_relations}
\title{Synonym table for various geographical scales}
\format{
A data frame with 3,142 rows and 6 variables:
\describe{
  \item{fips_state}{2-digit FIPS code indicating a state}
  \item{fips_county}{3-digit FIPS code indicating a county within a state}
  \item{fips_tract}{6-digit FIPS code indicating a tract within a county}
  \item{state_name_abbr}{2-character, capitalized state abbreviation}
  \item{state_name_full}{fully spelled and case-sensitive state name}
  \item{county_name_abbr}{abbreviated county name (usually minus the word 'County')}
  \item{county_name_full}{fully spelled and case-sensitive county name}
  \item{region}{fully spelled out and case-sensitive NHTSA region and constituent states}
  \item{region_abbr}{abbreviated NHTSA region (ne, mw, s, w)}
}
}
\source{
\url{https://www.census.gov/geographies/reference-files/2015/demo/popest/2015-fips.html}
}
\usage{
geo_relations
}
\description{
A dataset providing different ways to refer to states and counties.
}
\keyword{datasets}
