% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fars.R
\name{get_fars}
\alias{get_fars}
\title{Get FARS data}
\usage{
get_fars(years = 2020, states = NULL, save_dir = getwd())
}
\arguments{
\item{years}{Years to be downloaded, in yyyy (character or numeric formats),
currently limited to 2016-2020}

\item{states}{(Optional) States to keep. Leave as NULL to keep all states.
Can be specified as full state name (e.g. "Virginia"), abbreviation ("VA"),
or FIPS code (51).}

\item{save_dir}{Directory to store files}
}
\description{
This function is a wrapper for \code{download_fars} and \code{prep_fars}.
    It downloads raw files from NHTSA and produces analysis-ready data files.
}
\details{
Prepared data files are stored in \code{save_dir}/FARS data/prepared/
}
\examples{
get_fars(c("2019", "2020"))
\dontrun{
get_fars(2016:2020, "Virginia")
get_fars(2020, "NC")
}
}
\seealso{
\code{download_fars} \code{prep_fars}
}
