% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Generate filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year required}
}
\value{
Returns character vector contains filenume for specified \code{year}
}
\description{
This is the function to generate filename depending on \code{year}.
}
\examples{
make_filename(2015)

}
