% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character or integer vector specifying the year(s)
(YYYY format).}
}
\value{
This function returns a character string with the correct file name
   for the specified year.
}
\description{
This function makes a file name by entering a year. You can specify the year
by using the \code{year} argument.
}
\details{
First, the function turns the year input into a integer, creating a
   numerical vector. The function then returns a character vector containing
   a formatted combination of text and variable values to specific a valid
   data file name for the desired year.
}
\examples{
make_filename(2013)
make_filename(c(2014, 2015))

}
