% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Print file name for given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer or string of any format of a year value}
}
\value{
this function will return a string of a compressed .csv file name specific to the
       accidents recorded in the given year
}
\description{
This function outputs a specific .csv.bz2 file name based on a a given year. The
            The input can be in any format.
}
\examples{
\dontrun{
make_filename(year=2013)
make_filename(2014)
}

}
