% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cases.R
\name{motorcycle}
\alias{motorcycle}
\title{(Internal) Find crashes involving motorcycles}
\usage{
motorcycle(df)
}
\arguments{
\item{df}{The FARS or GESCRSS data object to be searched.}
}
\description{
These internal functions take the FARS object created by use_fars and look
    for various cases, such as distracted or drowsy drivers.
}
