% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_fars.R
\name{prep_fars}
\alias{prep_fars}
\title{Prepare downloaded FARS files for use}
\usage{
prep_fars(y, wd, rawfiles, prepared_dir, states)
}
\arguments{
\item{y}{year, to be passed from \code{prep_fars}}

\item{wd}{working directory, , to be passed from \code{prep_fars}}

\item{rawfiles}{dataframe translating filenames into standard terms,
to be passed from \code{prep_fars}}

\item{prepared_dir}{the location where prepared files will be saved,
to be passed from \code{prep_fars}}

\item{states}{(Optional) Inherits from get_fars()}
}
\value{
Produces six files: yyyy_flat.rds, yyyy_multi_acc.rds,
    yyyy_multi_veh.rds, yyyy_multi_per.rds, yyyy_events.rds, and codebook.rds
}
\description{
Prepare downloaded FARS files for use
}
