% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_basic_csv.R
\name{read_basic_csv}
\alias{read_basic_csv}
\title{(Internal) Takes care of basic CSV reading}
\usage{
read_basic_csv(x, wd, rawfiles)
}
\arguments{
\item{x}{The cleaned name of the data table (CSV).}

\item{wd}{The working directory for these files}

\item{rawfiles}{The data frame connecting raw filenames to cleaned ones.}
}
\description{
(Internal) Takes care of basic CSV reading
}
