% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_basic_sas.R
\name{read_basic_sas}
\alias{read_basic_sas}
\title{(Internal) Takes care of basic SAS file reading}
\usage{
read_basic_sas(x, wd, rawfiles, catfile = paste0(wd, "formats.sas7bcat"))
}
\arguments{
\item{x}{The cleaned name of the data table (SAS7BDAT).}

\item{wd}{The working directory for these files}

\item{rawfiles}{The data frame connecting raw filenames to cleaned ones.}

\item{catfile}{The location of the sas7bcat file}
}
\description{
(Internal) Takes care of basic SAS file reading
}
\seealso{
read_basic_sas_nocat
}
