% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_basic_sas_nocat.R
\name{read_basic_sas_nocat}
\alias{read_basic_sas_nocat}
\title{(Internal) Takes care of basic SAS file reading when the bcat file creates an issue}
\usage{
read_basic_sas_nocat(x, wd, rawfiles, imps = NULL, omits = NULL)
}
\arguments{
\item{x}{The cleaned name of the data table (SAS7BDAT).}

\item{wd}{The working directory for these files}

\item{rawfiles}{The data frame connecting raw filenames to cleaned ones.}

\item{imps}{A named list to be passed to use_imp(). Each item's name represents
the non-imputed variable name; the item itself represents the related imputed variable.}

\item{omits}{Character vector of columns to omit}
}
\description{
(Internal) Takes care of basic SAS file reading when the bcat file creates an issue
}
