% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_fars.R
\name{use_fars}
\alias{use_fars}
\title{(Internal) Use FARS data files}
\usage{
use_fars(dir, prepared_dir, cache)
}
\arguments{
\item{dir}{Inherits from get_fars().}

\item{prepared_dir}{Inherits from get_fars().}

\item{cache}{Inherits from get_fars().}
}
\value{
Returns an object of class 'FARS' which is a list of six tibbles:
    flat, multi_acc, multi_veh, multi_per, events, and codebook.
}
\description{
Compile multiple years of prepared FARS data.
}
