% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_fars.R
\name{use_fars}
\alias{use_fars}
\title{(Internal) Use FARS data files}
\usage{
use_fars(prepared_dir = "FARS data", years = NULL, states = NULL)
}
\arguments{
\item{prepared_dir}{Directory where prepared files are currently saved.}

\item{years}{(Optional) Years to keep.}

\item{states}{(Optional) States to keep.}
}
\value{
Returns an object of class 'FARS' which is a list of five tibbles:
    flat, multi_acc, multi_veh, multi_per, and events
}
\description{
Combine multiple years of prepared FARS data stored in CSV files and bring
    into the current environment.
}
